requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.12
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bleach==3.1.5
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clyent==1.2.2
colorama==0.4.3
conda==4.8.3
conda-build==3.18.9
conda-package-handling==1.7.0
conda-verify==3.4.2
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
future==0.18.2
glob2==0.7
gunicorn==20.0.4
idna==2.9
imageio==2.6.1
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython @ file:///C:/ci/ipython_1592330400575/work
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.4
jupyterlab-server==1.1.5
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.2
numpy==1.18.5
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12
PyQt5-sip==4.19.19
pyqtgraph==0.10.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
terminado==0.8.3
testpath==0.4.4
tk==0.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
